Ajit | Moved icons and images to static assets
