+++ packages/webapp/src/components/navbar/Navbar.tsx
@@ -5,24 +5,22 @@ import Typography from '@material-ui/core/Typography';
 import Avatar from "@material-ui/core/Avatar";
 import PersonIcon from '@material-ui/icons/Person';
 import { useHistory } from 'react-router'
-import HomeLogo from './Logo.png'
+import HomeLogo from '../../assets/Logo.png'
 
 import './Navbar.scss';
-import { IconButton } from "@material-ui/core";
+
 const Navbar = () => {
     const browserHistory = useHistory();
     return (
         <div className='navbar'>
             <AppBar position="static">
                 <Toolbar className="toolbar">
-                    <IconButton onClick={()=>{browserHistory.push('/')}}>
-                    <Avatar variant="rounded" src="Logo.png" className="navbar-logo">
-                    <img alt="Home Logog" src={HomeLogo} />
-                    </Avatar>
-                    </IconButton>
+                        <Avatar variant="rounded" src="Logo.png" className="navbar-logo" onClick={() => { browserHistory.push('/') }}>
+                            <img alt="Home Logog" src={HomeLogo} />
+                        </Avatar>
                     <Typography variant="h6" className='title'>The Retro App</Typography>
                     <Avatar className='user-icon'>
-                        <PersonIcon/>
+                        <PersonIcon />
                     </Avatar>
                 </Toolbar>
             </AppBar>
